Add files via upload
The file "Edusafe.py" can run in the platform called VScode, only when it is given with access "Run as administrator". The file renders the error if we directly use the code because the pic/image used differs with the location/path the code holds. And also make sure it should use the backslashes (\) in file paths on Windows.