LICENSE
Add MIT License

Include the MIT License to grant users permission to use, modify, and distribute the software for any purpose. The MIT License is a permissive open-source license that provides users with the freedom to use the software with minimal restrictions.

The decision to apply the MIT License was made to foster collaboration and encourage contributions from the open-source community. By choosing a permissive license, we aim to make our project accessible to a wide range of users and developers, while ensuring that proper credit and copyright notices are maintained.

For more information, see the LICENSE file.